+++ .moban.d/helper_and_partial.rst.jj2
@@ -4,13 +4,12 @@ Let's save the following file a `script.py` under `helper_and_partial` folder:
 
 .. code-block:: python
 
-   from moban_handlebars.helpers import HandlebarHelper
-   from moban_handlebars.partials import register_partial
+   from moban_handlebars.api import Helper, register_partial
 
    register_partial('header', '<h1>People</h1>')
 
 
-   @HandlebarHelper('list')
+   @Helper('list')
    def _list(this, options, items):
        result = [u'<ul>']
        for thing in items:
@@ -24,7 +23,7 @@ Let's save the following file a `script.py` under `helper_and_partial` folder:
 Let invoke handlebar template:
 {% raw %}
 
-.. code-block: bash
+.. code-block:: bash
 
    $ moban --template-type hbs -pd helper_and_partial -c data.json "{{>header}}{{#list people}}{{name}} {{age}}{{/list}}"
    Handlebars-ing {{>header}... to moban.output
